modules/api_locations_even/requirements.txt
grpcio == 1.39.0
grpcio-tools == 1.34.0
kafka-python == 2.0.2